requirements.txt
Pillow>=9.0.0
pillow_heif
opencv-python
imageio
scikit-image
tqdm